// src/service/request/type.ts
import { AxiosRequestConfig, AxiosResponse } from 'axios'

export interface XRequestInterceptors<T = AxiosResponse> {
  requestInterceptor?: (config: AxiosRequestConfig) => AxiosRequestConfig
  requestInterceptorCatch?: (error: any) => any
  responseInterceptor?: (res: T) => T
  responseInterceptorCatch?: (error: any) => any
}

export interface XRequestConfig<T = AxiosResponse> extends AxiosRequestConfig {
  interceptors?: XRequestInterceptors<T>
  showLoading?: boolean
}